+++ NEWS.md
@@ -1,3 +1,7 @@
+* New function `units()` that determines units for each variable,
+  using heuristics from the naming system.
+
+
 # ReboundTools 0.1.1 (2021-01-02)
 
 * Now at 246 tests, all passing.